Update AppCI324DIFF.lua
Support both Ci324PN and CI324 report

+++ AppCI324DIFF.lua
@@ -1,4 +1,4 @@
--- Diff Report : Find and calculate the difference between files CI324 FDS Monthly Trial Balance (CSV Format from Portal DWH)
+-- Diff Report : Find and calculate the difference between files CI324 PN FDS Monthly Trial Balance (CSV Format from Portal DWH)
 -- Interpreter : (Novan's Modified)lua.exe
 -- 10:03 PM Sunday, February 10, 2019 Jakarta City
 
@@ -12,8 +12,10 @@
 -- 7. 			Delta
 -- 8. 			Officer c17 c18
 			
-list_acc = {}
-OUTPUT_FILE = "CI324DIFF"
+local list_acc = {}
+local OUTPUT_FILE = "CI324PNDIFF"
+local HEADER_CI324 = "periode,branch,curr,currdesc,prodtype,acctno,serialno,princpamt,sname,wdrint,issuedt,matdt,rate,intdisp,textbox22"
+local HEADER_CI324_PN = 	 "periode,branch,curr,currdesc,prodtype,acctno,serialno,princpamt,sname,wdrint,issuedt,matdt,rate,intdisp,textbox22,CFAREF,SNAME_FO"
 
 function format_account(s)
     local rek_len = string.len(s)
@@ -24,25 +26,37 @@ function format_account(s)
     return s
 end
 
+function ReadRegistry(key, value)
+	local fi, data, content, data_type
+	
+	fi = io.popen(string.format('reg QUERY "%s" /v %s', key, value))
+	data = nil
+	if fi then
+		content = fi:read("*a")
+		data_type, data = content:match(value..'%s+(%S+)%s+(.+)\n\n')
+		fi:close()
+	end
+
+	return data_type, data
+end
+
 function format_number(v)
 	local s
-	local unary
+	local unary, sep_thousand
 	
 	if v < 0 then
-		--s = string.format("%d", math.floor(-v))
 		s = tostring(-v)
 		unary = "-"
 	else
-		--s = string.format("%d", math.floor(v))
 		s = tostring(v)
 		unary = ""
 	end
     
     local pos = string.len(s) % 3
-
+	data_type, thousand_sep = ReadRegistry('HKCU\\Control Panel\\International', 'sThousand')
+	if thousand_sep == nil then thousand_sep = '.' end
     if pos == 0 then pos = 3 end
-    return unary..string.sub(s, 1, pos).. string.gsub(string.sub(s, pos+1), "(...)", ".%1")
---	return v
+    return unary..string.sub(s, 1, pos).. string.gsub(string.sub(s, pos+1), "(...)", thousand_sep.."%1")
 end
 
 function FindFirstSeparator(line)
@@ -63,13 +77,35 @@ function FindFirstSeparator(line)
 	return sep
 end
 
-res, ReportFileName1, ReportFileName2, output_sep, limit_res = iup.GetParam("Pilih Report CI324 dalam Format CSV (Sumber: DWH)", nil, [=[
-Report Posisi Awal: %f[OPEN|*.csv;*.txt|CURRENT|NO|NO]\n
-Report Posisi Akhir: %f[OPEN|*.csv;*.txt|CURRENT|NO|NO]\n
-Output Separator: %l|,|;|\n
+function Report_Type(header)
+	header = header:gsub(string.char(0x0D),'')
+	header = header:gsub(string.char(0x0A),'')
+	if header == string.char(0xEF, 0xBB, 0xBF)..HEADER_CI324 then
+		return "CI324"
+	elseif header == string.char(0xEF, 0xBB, 0xBF)..HEADER_CI324_PN then
+		return "CI324PN"
+	elseif header == HEADER_CI324 then 
+		return "CI324"
+	elseif header == HEADER_CI324_PN then
+		return "CI324PN"
+	else
+		return nil
+	end
+end
+
+res, dummy, ReportFileName1, ReportFileName2, limit_res = iup.GetParam("Pilih Report CI324 PN dalam Format CSV (Sumber: DWH)", nil, [=[
+Sumber Data: %m\n
+Report Posisi Awal: %f[OPEN|*CI324*.csv;*CI324*.gz|CURRENT|NO|NO]\n
+Report Posisi Akhir: %f[OPEN|*CI324*.csv;*CI324*.gz|CURRENT|NO|NO]\n
 Limit Result: %l|10|20|30|50|100|\n
 ]=]
-,"","",0,1)
+,"1. Buka Aplikasi BRISIM (https://brisim.bri.co.id)\n2. Pilih: DWH Reports\n3. Pilih: Critical Report\n4. Pilih: Table\n5. Pilih CI324(PN) - FDS MONTHLY TRIAL BALANCE BY PRODUCT TYPE(1 ROW)\n6. Download dan Save dalam format CSV", "C:\\Lua\\data\\20201231 CI324Modif.csv.gz","C:\\Lua\\data\\20210131 CI324Modif.csv.gz",1)
+
+data_type, output_sep = ReadRegistry('HKCU\\Control Panel\\International', 'sList')
+data_type, decimal_sep = ReadRegistry('HKCU\\Control Panel\\International', 'sDecimal')
+if output_sep == decimal_sep then output_sep =';' end
+if output_sep == nil then output_sep = ',' end
+
 
 if ReportFileName1 == "" or ReportFileName2 == "" then
 	print("Please select two reports to be compared")
@@ -78,15 +114,25 @@ if ReportFileName1 == "" or ReportFileName2 == "" then
 end
 
 -- convert Unicode to ANSI
+if ReportFileName1:match('%.gz$') == nil then
 print('Converting '..ReportFileName1..' to ANSI encoding')
 os.execute('type "'..ReportFileName1..'" > '..'tmp.csv')
 os.remove(ReportFileName1)
 os.rename('tmp.csv', ReportFileName1)
+f_lines1 = io.lines
+else
+f_lines1 = gzio.lines
+end
 
+if ReportFileName1:match('%.gz$') == nil then
 print('Converting '..ReportFileName2..' to ANSI encoding')
 os.execute('type "'..ReportFileName2..'" > '..'tmp.csv')
 os.remove(ReportFileName2)
 os.rename('tmp.csv', ReportFileName2)
+f_lines2 = io.lines
+else
+f_lines2 = gzio.lines
+end
 
 -- Load first data into table list_acc
 t1 = os.clock()
@@ -94,10 +140,17 @@ print('Loading data from '..ReportFileName1)
 no = 1
 sep = ','
 posisi_report1 = ''
-for line in io.lines(ReportFileName1) do
+report1_type = ''
+report2_type = ''
+for line in f_lines1(ReportFileName1) do
 	-- process header
 	if no == 1 then
 		sep = FindFirstSeparator(line)
+		report1_type = Report_Type(line)
+		if report1_type ~= "CI324PN" and report1_type ~= "CI324" then
+			iup.Message("Error","Report [Awal] yang dipilih bukan Report \"CI324PN\" atau \"CI324\" dalam format CSV.\nSilahkan download ulang dari BRISIM atau \npilih kembali report yang sesuai.")
+			return -1
+		end
 	else
 		-- process data
 		f = csv.parse(line, sep)
@@ -107,7 +160,11 @@ for line in io.lines(ReportFileName1) do
 			acc_type_tenor = f[5].." - "..f[14]
 			acc_maturity = f[12]
 			acc_name = f[9]
-			acc_officer = f[16].." - "..f[17]
+			if report1_type == "CI324" then
+				acc_officer = ''
+			else
+				acc_officer = f[16]..' - '..f[17]
+			end
 			acc_balance = string.gsub(string.sub(f[8], 1, #f[8]-3), ",", "")
 			acc_balance = tonumber(acc_balance)
 			list_acc[acc_no] = {acc_type_tenor, acc_maturity, acc_name, acc_balance, 0, -acc_balance, acc_officer}
@@ -121,10 +178,17 @@ print('Loading data from '..ReportFileName2)
 no = 1
 sep = ','
 posisi_report2 = ''
-for line in io.lines(ReportFileName2) do
+fo = io.open(OUTPUT_FILE.."_NEW.csv", "w")
+fo:write('Rekening'..output_sep..'Tipe'..output_sep..'Nama'..output_sep..'Tanggal Buka'..output_sep..'Jatuh Tempo'..output_sep..'Rate'..output_sep..'Tenor'..output_sep..'Rollover'..output_sep..'Currency'..output_sep..'Pokok'..output_sep..'PN_Pengelola\n')
+for line in f_lines2(ReportFileName2) do
 	-- only process line begin with number, skipping header
 	if no == 1 then
 		sep = FindFirstSeparator(line)
+		report2_type = Report_Type(line)
+		if report2_type ~= "CI324PN" and report1_type ~= "CI324" then
+			iup.Message("Error","Report [Awal] yang dipilih bukan Report \"CI324PN\" atau \"CI324\" dalam format CSV.\nSilahkan download ulang dari BRISIM atau \npilih kembali report yang sesuai.")
+			return -1
+		end
 	else
 		f = csv.parse(line, sep)
 		if f[1] ~= "" then
@@ -133,7 +197,11 @@ for line in io.lines(ReportFileName2) do
 			acc_type_tenor = f[5].." - "..f[14]
 			acc_maturity = f[12]
 			acc_name = f[9]
-			acc_officer = f[16].." - "..f[17]
+			if report2_type == "CI324" then
+				acc_officer = ''
+			else
+				acc_officer = f[16]..' - '..f[17]
+			end
 			acc_balance = string.gsub(string.sub(f[8], 1, #f[8]-3), ",", "")
 			acc_balance = tonumber(acc_balance)
 			if list_acc[acc_no] then
@@ -141,11 +209,26 @@ for line in io.lines(ReportFileName2) do
 				list_acc[acc_no][6] = list_acc[acc_no][5] - list_acc[acc_no][4]
 			else
 				list_acc[acc_no] = {acc_type_tenor, acc_maturity, acc_name, 0, acc_balance, acc_balance, acc_officer}
+				issuedt = csv.parse(f[11],'/')
+				matdt = csv.parse(f[12],'/')
+				fo:write(string.format('%s%s%s%s"%s"%s%s/%s/%s%s%s/%s/%s%s"%s"%s%s%s%s%s%s%s"%s"%s%s\n', 
+					format_account(acc_no), output_sep,
+					f[5], output_sep,
+					acc_name, output_sep,
+					issuedt[1], issuedt[2], issuedt[3], output_sep,
+					matdt[1], matdt[2], matdt[3], output_sep, 
+					f[13]:gsub('%.',decimal_sep), output_sep,
+					f[14], output_sep,
+					f[15], output_sep,
+					f[3], output_sep,
+					format_number(acc_balance), output_sep,
+					acc_officer))
 			end
 		end
 	end
 	no = no + 1
 end
+fo:close()
 
 print('Sorting descending')
 sorted_list_acc = {}
@@ -175,14 +258,14 @@ end
 ii = 0
 fo = io.open(OUTPUT_FILE..".csv", "w")
 fo2 = io.open(OUTPUT_FILE..".htm", "w")
-fo:write('No Rek'..output_sep..'CIF'..output_sep..'Nama'..output_sep..'Posisi Awal'..output_sep..'Posisi Akhir'..output_sep..'Delta\n')
+fo:write('No Rek'..output_sep..'CIF'..output_sep..'Nama'..output_sep..'Posisi Awal'..output_sep..'Posisi Akhir'..output_sep..'Delta'..output_sep..'Officer\n')
 fo2:write([=[
 <!DOCTYPE html>
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
  
-<title>BRI Reporting Tool: Delta Tabungan CI324</title>
+<title>BRI Reporting Tool: Delta Deposito CI324</title>
 <style type="text/css"> 
 body, html  { height: 100%; }
 html, body, div, span, applet, object, iframe,
@@ -355,6 +438,7 @@ fo2:close()
 fo:close()
 print('=== Done in '..(os.clock()-t1)..' ===')
 os.execute(OUTPUT_FILE..".htm")
---os.execute("pause")
 
-
+if iup.Alarm("Open List of New Created Account", "Buka file daftar rekening baru yang dibuat \ndalam periode "..posisi_report1.." sampai "..posisi_report2.." ? " ,"Ya" ,"Tidak") == 1 then
+	os.execute(OUTPUT_FILE.."_NEW.csv")
+end
